:recycle: refactor: Query Service 위치 변경
